:recycle: Move the mixer crickets proving logic to `proof/mixer.rs`
